backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
SQLAlchemy>=2.0
asyncpg>=0.29
alembic>=1.13
python-dotenv>=1.0
pydantic[email]
passlib[argon2]>=1.7
PyJWT>=2.8


